Added fixes discussed
