+++ client/src/Common/Enums/VehicleCategoryEnum.ts
@@ -1,0 +1,9 @@
+enum VehicleCategories {
+    NONE,
+    MOTORBIKE,
+    CAR,
+    TRUCK,
+    BICYCLE,
+}
+
+export default VehicleCategories;
