--- client/src/Common/Enums/VehicleCatehoryEnum.ts
@@ -1,9 +1,0 @@
-enum VehicleCategories {
-    NONE,
-    MOTORBIKE,
-    CAR,
-    TRUCK,
-    BICYCLE,
-}
-
-export default VehicleCategories;
