--- client/src/Helpers/Defaults.tsx
@@ -1,76 +1,0 @@
-import {
-    IVehicleInfo,
-    IOwner,
-    IMemberAttributes,
-    IPostAttributes,
-    IComment,
-    IData,
-    IInputFields,
-    IChip,
-} from '../Common/Interfaces/interfaces';
-
-export const defaultVehicleData: IVehicleInfo = {
-    vehicle_id: -1,
-    date_added: '',
-    description: '',
-    features: [],
-    images: [''],
-    location: '',
-    number_plate: '',
-    owner_id: '',
-    status: -1,
-    make: '',
-    model: '',
-    vin: '',
-    category: '',
-};
-
-export const defaultMemberAttributes: IMemberAttributes = {
-    display_name: '',
-    profile_image: '',
-};
-
-export const defaultPostAttributes: IPostAttributes = {
-    message: '',
-    confirmation_image: '',
-    active_state: false,
-};
-
-export const defaultOwner: IOwner = {
-    member_attributes: defaultMemberAttributes,
-};
-
-export const defaultComment: IComment = {
-    post_id: -1,
-    type: -1,
-    date_added: '',
-    member_attributes: defaultMemberAttributes,
-    post_attributes: defaultPostAttributes,
-};
-
-export const defaultData: IData = {
-    vehicle: defaultVehicleData,
-    owner: defaultOwner,
-    posts: [defaultComment],
-};
-
-export const defaultChip: IChip = {
-    key: -1,
-    value: '',
-};
-
-export const defaultInputs: IInputFields = {
-    numberPlate: '',
-    vin: '',
-    make: '',
-    model: '',
-    category: 0,
-    primaryColour: '',
-    secondaryColour: '',
-    dateStolen: new Date(),
-    location: '',
-    features: '',
-    featuresArray: [],
-    description: '',
-    v5cVerificationDate: '',
-};
